+++ dsl/src/test/scala/com/crobox/clickhouse/DslTestSpec.scala
@@ -1,5 +1,6 @@
 package com.crobox.clickhouse
 
+import com.crobox.clickhouse.dsl.{InternalQuery, OperationalQuery, TableColumn}
 import com.crobox.clickhouse.dsl.language.ClickhouseTokenizerModule
 import com.crobox.clickhouse.testkit.ClickhouseSpec
 import com.typesafe.config.{Config, ConfigFactory}
@@ -15,8 +16,26 @@ trait DslTestSpec
     with Matchers
     with ClickhouseSpec
     with TestSchema
-    with ClickhouseSQLSupport
     with ClickhouseTokenizerModule {
 
   override val config: Config = ConfigFactory.load()
+
+  def toSQL(condition: TableColumn[Boolean]): String = toSQL(Option(condition))
+
+  def toSQL(condition: Option[TableColumn[Boolean]]): String = {
+    val s = toSql(InternalQuery(where = condition))
+    s.substring("WHERE ".length, s.indexOf("FORMAT")).trim
+  }
+
+  def toSQL(query: OperationalQuery): String = toSQL(query, stripBeforeWhere = true)
+
+  def toSQL(query: OperationalQuery, stripBeforeWhere: Boolean): String = {
+    val sql = toSql(query.internalQuery)
+    if (stripBeforeWhere) {
+      sql.indexOf("WHERE") match {
+        case idx if idx > 0 => sql.substring(idx, sql.indexOf(" FORMAT")).trim
+        case _              => sql.substring(0, sql.indexOf(" FORMAT")).trim
+      }
+    } else sql.substring(0, sql.indexOf(" FORMAT")).trim
+  }
 }
